Backend/requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.2.0
CacheControl==0.13.1
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.1
distro==1.9.0
dnspython==2.5.0
exceptiongroup==1.2.0
fastapi==0.109.0
firebase-admin==6.4.0
frozenlist==1.4.1
google-ai-generativelanguage==0.4.0
google-api-core==2.15.0
google-api-python-client==2.115.0
google-auth==2.27.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.14.0
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-generativeai==0.3.2
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
googlemaps==4.10.0
googletrans==2.4.0
greenlet==3.0.3
grpcio==1.60.0
grpcio-status==1.60.0
h11==0.12.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.1.5
httpcore==0.15.0
httplib2==0.22.0
httptools==0.6.1
httpx==0.24.1
hyperframe==5.2.0
idna==3.6
msgpack==1.0.7
multidict==6.0.4
openai==1.2.0
proto-plus==1.23.0
protobuf==4.25.2
psycopg2==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
PyJWT==2.8.0
pymongo==4.6.1
pyparsing==3.1.1
python-dotenv==1.0.1
PyYAML==6.0.1
requests==2.31.0
rfc3986==1.5.0
rsa==4.9
sniffio==1.3.0
SQLAlchemy==2.0.25
starlette==0.35.1
tqdm==4.66.1
typing_extensions==4.9.0
uritemplate==4.1.1
urllib3==2.1.0
uvicorn==0.27.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
yarl==1.9.4
